Refactorizar routes y utils
