+++ utils/index.js
@@ -1,0 +1,2 @@
+const generateRandomString = require('./generateRandomString');
+module.exports = { generateRandomString };
